+++ pyvista/utilities/__init__.py
@@ -1,5 +1,4 @@
 """Utilities routines."""
-from .common import perlin_noise, sample_function
 from .errors import (
     GPUInfo,
     Observer,
@@ -73,3 +72,5 @@ from .reader import (
     XMLMultiBlockDataReader,
     DICOMReader,
 )
+
+# from .common import perlin_noise, sample_function
